+++ R/make_images.R
@@ -1,8 +1,7 @@
-#' @importFrom grDevices bmp cairo_pdf cairo_ps dev.off jpeg png svg tiff
 #' @import grid
 
 COMPONENT_AND_SIDES <- c("tile_back", "tile_face", "coin_back", "coin_face",
-           "die_face", "suitdie_face", 
+           "die_face", "suitdie_face",
            "pawn_face", "pawn_back", "belt_face",  "saucer_face", "saucer_back",
            "pyramid_face", "pyramid_back", "pyramid_left", "pyramid_right", "pyramid_top",
            "matchstick_face", "matchstick_back")
@@ -12,7 +11,7 @@ has_suit <- function(cs) {
 }
 has_rank <- function(cs) {
     !(cs %in% c("tile_back", "coin_back", "suitdie_face",
-                "pawn_face", "pawn_back", "belt_face", 
+                "pawn_face", "pawn_back", "belt_face",
                 "saucer_face", "saucer_back"))
 }
 
@@ -47,19 +46,21 @@ pp_device <- function(filename, piece_side=NULL, cfg=list(), angle=0, suit=1, ra
     }
     bg <- "transparent"
     switch(format,
-           bmp = bmp(filename, width, height, "in", res=res, bg=bg),
-           jpeg = jpeg(filename, width, height, "in", res=res, bg=bg),
-           pdf = cairo_pdf(filename, width, height, bg=bg),
-           png = png(filename, width, height, "in", res=res, bg=bg),
-           ps = cairo_ps(filename, width, height, bg=bg),
-           svg = svg(filename, width, height, bg=bg),
-           tiff = tiff(filename, width, height, "in", res=res, bg=bg))
+           bmp = grDevices::bmp(filename, width, height, "in", res=res, bg=bg),
+           jpeg = grDevices::jpeg(filename, width, height, "in", res=res, bg=bg),
+           jpg = grDevices::jpeg(filename, width, height, "in", res=res, bg=bg),
+           pdf = grDevices::cairo_pdf(filename, width, height, bg=bg),
+           png = grDevices::png(filename, width, height, "in", res=res, bg=bg),
+           ps = grDevices::cairo_ps(filename, width, height, bg=bg),
+           svg = grDevices::svg(filename, width, height, bg=bg),
+           svgz = grDevices::svg(filename, width, height, bg=bg),
+           tiff = grDevices::tiff(filename, width, height, "in", res=res, bg=bg))
     pushViewport(viewport(angle=angle, name="main"))
 }
 
-piece_filename <- function(directory, cfg, piece_side, format, angle, 
+piece_filename <- function(directory, piece_side, format, angle,
                                suit=NULL, rank=NULL) {
-    filename <- paste0(piece_side, 
+    filename <- paste0(piece_side,
                        ifelse(is.null(suit), "", paste0("_s", suit)),
                        ifelse(is.null(rank), "", paste0("_r", rank)),
                        paste0("_t", angle), paste0(".", format))
@@ -82,46 +83,63 @@ piece_filename <- function(directory, cfg, piece_side, format, angle,
 #'         save_piece_images(cfg, directory=tempdir(), format="png", angle=90)
 #'     }
 #'   }
-#'   
+#'
 #' @export
 save_piece_images <- function(cfg=pp_cfg(), directory=tempdir(), format="svg", angle=0) {
+    current_dev <- grDevices::dev.cur()
+    if (current_dev > 1) on.exit(grDevices::dev.set(current_dev))
+
+    if (!dir.exists(directory)) stop(paste("directory", directory, "does not exist"))
     for (f in format) {
         for (a in angle) make_images_helper(directory, cfg, f, a)
     }
 }
 
+pp_dev_off <- function(f, format) {
+    v <- grDevices::dev.off()
+    if (format == "svgz") {
+        rcon <- file(f, "r")
+        svg <- readLines(rcon)
+        close(rcon)
+        wcon <- gzfile(f, "w")
+        writeLines(svg, wcon)
+        close(wcon)
+    }
+    invisible(v)
+}
+
 make_images_helper <- function(directory, cfg, format, angle) {
     suppressWarnings({
         for (cs in COMPONENT_AND_SIDES) {
-            if(!has_suit(cs) && !has_rank(cs)) {
-                f <- piece_filename(directory, cfg, cs, format, angle)
+            if (!has_suit(cs) && !has_rank(cs)) {
+                f <- piece_filename(directory, cs, format, angle)
                 pp_device(f, cs, cfg, angle)
                 grid.piece(cs, NA, NA, cfg)
-                invisible(dev.off())
+                pp_dev_off(f, format)
             }
-            if(has_suit(cs) && !has_rank(cs)) {
+            if (has_suit(cs) && !has_rank(cs)) {
                 for (suit in 1:get_n_suits(cfg)) {
-                    f <- piece_filename(directory, cfg, cs, format, angle, suit)
+                    f <- piece_filename(directory, cs, format, angle, suit)
                     pp_device(f, cs, cfg, angle)
                     grid.piece(cs, suit, NA, cfg)
-                    invisible(dev.off())
+                    pp_dev_off(f, format)
                 }
             }
-            if(!has_suit(cs) && has_rank(cs)) {
+            if (!has_suit(cs) && has_rank(cs)) {
                 for (rank in 1:get_n_ranks(cfg)) {
-                    f <- piece_filename(directory, cfg, cs, format, angle, rank=rank)
+                    f <- piece_filename(directory, cs, format, angle, rank=rank)
                     pp_device(f, cs, cfg, angle)
                     grid.piece(cs, NA, rank, cfg)
-                    invisible(dev.off())
+                    pp_dev_off(f, format)
                 }
             }
-            if(has_suit(cs) && has_rank(cs)) {
+            if (has_suit(cs) && has_rank(cs)) {
                 for (suit in 1:get_n_suits(cfg)) {
                     for (rank in 1:get_n_ranks(cfg)) {
-                        f <- piece_filename(directory, cfg, cs, format, angle, suit, rank)
+                        f <- piece_filename(directory, cs, format, angle, suit, rank)
                         pp_device(f, cs, cfg, angle, rank=rank)
                         grid.piece(cs, suit, rank, cfg)
-                        invisible(dev.off())
+                        pp_dev_off(f, format)
                     }
                 }
             }